+++ server/api/controllers/transaction.js
@@ -128,9 +128,135 @@ const adminCheckTransaction = (req, res) => {
 		});
 };
 
+const getAdminDeposits = (req, res) => {
+	loggerDeposits.verbose(
+		req.uuid,
+		'controllers/admin/getDeposits/auth',
+		req.auth
+	);
+
+	const { user_id, currency, limit, page, order_by, order, start_date, end_date, status, dismissed, rejected, processing, waiting, format } = req.swagger.params;
+
+	toolsLib.users.getUserDepositsByKitId(user_id.value, currency.value, status.value, dismissed.value, rejected.value, processing.value, waiting.value, limit.value, page.value, order_by.value, order.value, start_date.value, end_date.value, format.value)
+		.then((data) => {
+			if (format.value) {
+				if (data.data.length === 0) {
+					throw new Error('No data found');
+				}
+				res.setHeader('Content-disposition', `attachment; filename=${toolsLib.getKitConfig().api_name}-users-deposits.csv`);
+				res.set('Content-Type', 'text/csv');
+				return res.status(202).send(data);
+			} else {
+				return res.json(data);
+			}
+		})
+		.catch((err) => {
+			loggerDeposits.error(
+				req.uuid,
+				'controllers/admin/getDeposits',
+				err.message
+			);
+			return res.status(err.status || 400).json({ message: err.message });
+		});
+};
+
+const getAdminWithdrawals = (req, res) => {
+	loggerDeposits.verbose(
+		req.uuid,
+		'controllers/admin/getWithdrawals/auth',
+		req.auth
+	);
+
+	const { user_id, currency, limit, page, order_by, order, start_date, end_date, status, dismissed, rejected, processing, waiting, format } = req.swagger.params;
+
+	toolsLib.users.getUserWithdrawalsByKitId(user_id.value, currency.value, status.value, dismissed.value, rejected.value, processing.value, waiting.value, limit.value, page.value, order_by.value, order.value, start_date.value, end_date.value, format.value)
+		.then((data) => {
+			if (format.value) {
+				if (data.data.length === 0) {
+					throw new Error('No data found');
+				}
+				res.setHeader('Content-disposition', `attachment; filename=${toolsLib.getKitConfig().api_name}-users-deposits.csv`);
+				res.set('Content-Type', 'text/csv');
+				return res.status(202).send(data);
+			} else {
+				return res.json(data);
+			}
+		})
+		.catch((err) => {
+			loggerDeposits.error(
+				req.uuid,
+				'controllers/admin/getWithdrawals',
+				err.message
+			);
+			return res.status(err.status || 400).json({ message: err.message });
+		});
+};
+
+const getUserDeposits = (req, res) => {
+	loggerDeposits.verbose(
+		req.uuid,
+		'controllers/user/getUserDeposits auth',
+		req.auth.sub
+	);
+	const user_id = req.auth.sub.id;
+	const currency = req.swagger.params.currency.value || '';
+	const { limit, page, order_by, order, start_date, end_date, format } = req.swagger.params;
+
+	toolsLib.users.getUserDepositsByKitId(user_id, currency, limit.value, page.value, order_by.value, order.value, start_date.value, end_date.value, format.value)
+		.then((data) => {
+			if (format.value) {
+				if (data.data.length === 0) {
+					throw new Error('No data found');
+				}
+				res.setHeader('Content-disposition', `attachment; filename=${toolsLib.getKitConfig().api_name}-deposits.csv`);
+				res.set('Content-Type', 'text/csv');
+				return res.status(202).send(data);
+			} else {
+				return res.json(data);
+			}
+		})
+		.catch((err) => {
+			loggerDeposits.error('controllers/user/getUserDeposits', err.message);
+			return res.status(err.status || 400).json({ message: err.message });
+		});
+};
+
+const getUserWithdrawals = (req, res) => {
+	loggerDeposits.verbose(
+		req.uuid,
+		'controllers/user/getUserWithdrawals auth',
+		req.auth.sub
+	);
+	const user_id = req.auth.sub.id;
+	const currency = req.swagger.params.currency.value || '';
+	const { limit, page, order_by, order, start_date, end_date, format } = req.swagger.params;
+
+	toolsLib.users.getUserWithdrawalsByKitId(user_id, currency, limit.value, page.value, order_by.value, order.value, start_date.value, end_date.value, format.value)
+		.then((data) => {
+			if (format.value) {
+				if (data.data.length === 0) {
+					throw new Error('No data found');
+				}
+				res.setHeader('Content-disposition', `attachment; filename=${toolsLib.getKitConfig().api_name}-withdrawals.csv`);
+				res.set('Content-Type', 'text/csv');
+				return res.status(202).send(data);
+			} else {
+				return res.json(data);
+			}
+		})
+		.catch((err) => {
+			loggerDeposits.error('controllers/user/getUserWithdrawals', err.message);
+			return res.status(err.status || 400).json({ message: err.message });
+		});
+};
+
 module.exports = {
 	requestWithdrawal,
 	getWithdrawalFee,
 	adminCheckTransaction,
-	performWithdrawal
+	performWithdrawal,
+	getAdminDeposits,
+	getAdminWithdrawals,
+	getUserDeposits,
+	getUserWithdrawals
 };
