+++ server/api/controllers/user.js
@@ -449,96 +449,6 @@ const deleteHmacToken = (req, res) => {
 		});
 };
 
-const getUserTrades = (req, res) => {
-	loggerUser.verbose(
-		req.uuid,
-		'controllers/user/getUserTrades auth',
-		req.auth.sub
-	);
-
-	const user_id = req.auth.sub.id;
-	const { limit, page, order_by, order, start_date, end_date, format } = req.swagger.params;
-	const symbol = req.swagger.params.symbol.value;
-
-	if (symbol && !toolsLib.subscribedToPair(symbol)) {
-		loggerUser.error(req.uuid, 'controllers/user/getUserTrades', 'Invalid symbol');
-		return res.status(400).json({ message: 'Invalid symbol' });
-	}
-
-	toolsLib.order.getAllUserTradesNetworkByKidId(user_id, symbol, limit.value, page.value, order_by.value, order.value, start_date.value, end_date.value, format.value)
-		.then((data) => {
-			if (format.value) {
-				res.setHeader('Content-disposition', `attachment; filename=${toolsLib.getKitConfig().api_name}-trades.csv`);
-				res.set('Content-Type', 'text/csv');
-				return res.status(202).send(data);
-			} else {
-				return res.json(data);
-			}
-		})
-		.catch((err) => {
-			loggerUser.error(req.uuid, 'controllers/user/getUserTrades', err.message);
-			return res.status(err.status || 400).json({ message: err.message });
-		});
-};
-
-const getUserDeposits = (req, res) => {
-	loggerUser.verbose(
-		req.uuid,
-		'controllers/user/getUserDeposits auth',
-		req.auth.sub
-	);
-	const user_id = req.auth.sub.id;
-	const currency = req.swagger.params.currency.value || '';
-	const { limit, page, order_by, order, start_date, end_date, format } = req.swagger.params;
-
-	toolsLib.users.getUserDepositsByKitId(user_id, currency, limit.value, page.value, order_by.value, order.value, start_date.value, end_date.value, format.value)
-		.then((data) => {
-			if (format.value) {
-				if (data.data.length === 0) {
-					throw new Error('No data found');
-				}
-				res.setHeader('Content-disposition', `attachment; filename=${toolsLib.getKitConfig().api_name}-deposits.csv`);
-				res.set('Content-Type', 'text/csv');
-				return res.status(202).send(data);
-			} else {
-				return res.json(data);
-			}
-		})
-		.catch((err) => {
-			loggerUser.error('controllers/user/getUserDeposits', err.message);
-			return res.status(err.status || 400).json({ message: err.message });
-		});
-};
-
-const getUserWithdrawals = (req, res) => {
-	loggerUser.verbose(
-		req.uuid,
-		'controllers/user/getUserWithdrawals auth',
-		req.auth.sub
-	);
-	const user_id = req.auth.sub.id;
-	const currency = req.swagger.params.currency.value || '';
-	const { limit, page, order_by, order, start_date, end_date, format } = req.swagger.params;
-
-	toolsLib.users.getUserWithdrawalsByKitId(user_id, currency, limit.value, page.value, order_by.value, order.value, start_date.value, end_date.value, format.value)
-		.then((data) => {
-			if (format.value) {
-				if (data.data.length === 0) {
-					throw new Error('No data found');
-				}
-				res.setHeader('Content-disposition', `attachment; filename=${toolsLib.getKitConfig().api_name}-withdrawals.csv`);
-				res.set('Content-Type', 'text/csv');
-				return res.status(202).send(data);
-			} else {
-				return res.json(data);
-			}
-		})
-		.catch((err) => {
-			loggerUser.error('controllers/user/getUserWithdrawals', err.message);
-			return res.status(err.status || 400).json({ message: err.message });
-		});
-};
-
 const getUserStats = (req, res) => {
 	loggerUser.verbose(
 		req.uuid,
@@ -601,9 +511,6 @@ module.exports = {
 	getHmacToken,
 	createHmacToken,
 	deleteHmacToken,
-	getUserTrades,
-	getUserDeposits,
-	getUserWithdrawals,
 	getUserStats,
 	cancelWithdrawal
 };
